% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.add_contour_lines}
\alias{.add_contour_lines}
\title{Add Contour Lines to a ggplot2 Plot
Wrapper around \code{geom_contour} with preset defaults for colour, alpha, width.
Uses the correct linewidth/size argument depending on ggplot2 version.}
\usage{
.add_contour_lines(
  mapping = NULL,
  data = NULL,
  breaks = NULL,
  colour = "white",
  alpha = 0.3,
  width = 0.2,
  bins = NULL
)
}
\arguments{
\item{mapping}{Mapping aesthetic.}

\item{data}{Data frame.}

\item{breaks}{Break points for contours.}

\item{colour}{Colour of contour lines.}

\item{alpha}{Transparency level.}

\item{width}{Line width.}

\item{bins}{Number of bins for contour fill.}
}
\value{
A ggplot2 layer adding contour lines.
}
\description{
Add Contour Lines to a ggplot2 Plot
Wrapper around \code{geom_contour} with preset defaults for colour, alpha, width.
Uses the correct linewidth/size argument depending on ggplot2 version.
}
\keyword{internal}
