% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.geom_line_lw}
\alias{.geom_line_lw}
\title{Create a ggplot2 Line Layer with Version-Compatible Width
Creates a \code{geom_line} with a width argument adapted to ggplot2 version.}
\usage{
.geom_line_lw(mapping = NULL, data = NULL, ..., width = 1)
}
\arguments{
\item{mapping}{Mapping aesthetic.}

\item{data}{Data frame.}

\item{...}{Additional parameters passed to \code{geom_line}.}

\item{width}{Numeric line width for lines, default 1.}
}
\value{
ggplot2 layer for lines.
}
\description{
Create a ggplot2 Line Layer with Version-Compatible Width
Creates a \code{geom_line} with a width argument adapted to ggplot2 version.
}
\keyword{internal}
