% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.scale_fill_viridis_discrete}
\alias{.scale_fill_viridis_discrete}
\title{Scale Fill for Viridis Discrete Data}
\usage{
.scale_fill_viridis_discrete(name = "Assurance")
}
\arguments{
\item{name}{Character legend title (default "Assurance")}
}
\value{
ggplot2 fill scale object
}
\description{
Scale Fill for Viridis Discrete Data
}
\keyword{internal}
