% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-analysis.R
\name{print.powerbrmsINLA_seq_design}
\alias{print.powerbrmsINLA_seq_design}
\title{Print method for sequential design objects}
\usage{
\method{print}{powerbrmsINLA_seq_design}(x, ...)
}
\arguments{
\item{x}{A \code{"powerbrmsINLA_seq_design"} object.}

\item{...}{Unused; present for S3 compatibility.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for sequential design objects
}
